Convert signin-singup to typescript
